Added service to calc wall length
